pyproject.toml
[project]
name = "ugsys-identity-manager"
version = "0.1.0"
description = "Identity & access management service for the ugsys platform"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic[email]>=2.7",
    "email-validator>=2.1",
    "pydantic-settings>=2.3",
    "boto3>=1.34",
    "aioboto3>=13.0",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "python-multipart>=0.0.9",
    "structlog>=24.1",
    "mangum>=0.17",
    "aws-xray-sdk>=2.14",
    "ugsys-auth-client",
    "ugsys-logging-lib",
    "ugsys-event-lib",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "moto[dynamodb]>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "boto3-stubs[dynamodb]>=1.34",
    "bandit[toml]>=1.7",
    "safety>=3.2",
]

[tool.uv.sources]
ugsys-auth-client = { git = "https://github.com/awscbba/ugsys-shared-libs", tag = "auth-client/v0.1.0", subdirectory = "auth-client" }
ugsys-logging-lib  = { git = "https://github.com/awscbba/ugsys-shared-libs", tag = "logging-lib/v0.1.0",  subdirectory = "logging-lib" }
ugsys-event-lib    = { git = "https://github.com/awscbba/ugsys-shared-libs", tag = "event-lib/v0.1.0",    subdirectory = "event-lib" }

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "S", "ANN", "RUF"]
ignore = [
    "S101",    # assert in tests is fine
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106", "ANN"]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]  # assert in tests

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["src"]
omit = [
    "src/main.py",                              # composition root — wired at runtime
    "src/infrastructure/persistence/*",         # DynamoDB — covered by integration tests
    "src/infrastructure/tracing.py",            # X-Ray — requires Lambda runtime
    "src/infrastructure/logging.py",            # logging config — trivial setup
    "src/presentation/api/v1/auth.py",          # router — covered by integration tests
    "src/presentation/api/v1/users.py",         # router — covered by integration tests
    "src/presentation/api/v1/health.py",        # router — trivial
]

[tool.coverage.report]
fail_under = 80
show_missing = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_ignores = false

[[tool.mypy.overrides]]
module = [
    "jose",
    "jose.*",
    "aioboto3",
    "aioboto3.*",
    "aiobotocore",
    "aiobotocore.*",
    "aws_xray_sdk",
    "aws_xray_sdk.*",
    "moto",
    "moto.*",
    "mangum",
    "mangum.*",
]
ignore_missing_imports = true
